unify logging of exceptions in API
