+++ backend/src/core/auth/authz_user.py
@@ -1,16 +1,18 @@
 from typing import NoReturn
 
-from fastapi import Depends, Request
+from fastapi import Depends, Request, status
 from loguru import logger
 from sqlalchemy.orm import Session
 
 from common.crud_enum import Crud
 from common.dependencies import get_current_user, get_db_session
+from common.exception_handler import exception_handler
 from core.user.user_orm import UserORM
 from repos.db.crud_base import NoSuchElementError
 from repos.db.orm_base import ORMBase
 
 
+@exception_handler(status.HTTP_403_FORBIDDEN)
 class ForbiddenError(Exception):
     def __init__(self):
         super().__init__("User is not authorized")
