+++ backend/src/core/doc/source_document_crud.py
@@ -1,8 +1,10 @@
+from fastapi import status
 from fastapi.encoders import jsonable_encoder
 from sqlalchemy import func
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
+from common.exception_handler import exception_handler
 from common.sdoc_status_enum import SDocStatus
 from core.annotation.annotation_document_orm import AnnotationDocumentORM
 from core.doc.folder_crud import crud_folder
@@ -22,6 +24,7 @@ from repos.filesystem_repo import FilesystemRepo
 from systems.event_system.events import source_document_deleted
 
 
+@exception_handler(status.HTTP_500_INTERNAL_SERVER_ERROR)
 class SourceDocumentPreprocessingUnfinishedError(Exception):
     def __init__(self, sdoc_id: int):
         super().__init__(f"SourceDocument {sdoc_id} is still getting preprocessed!")
