--- backend/src/modules/crawler/crawler_exceptions.py
@@ -1,3 +1,0 @@
-class NoDataToCrawlError(Exception):
-    def __init__(self, what_msg: str):
-        super().__init__(what_msg)
