+++ backend/src/modules/eximport/codes/export_codes.py
@@ -10,7 +10,7 @@ from modules.eximport.codes.code_export_schema import (
     CodeExportCollection,
     CodeExportSchema,
 )
-from modules.eximport.no_data_export_error import NoDataToExportError
+from modules.eximport.export_exceptions import NoDataToExportError
 from repos.filesystem_repo import FilesystemRepo
