+++ backend/src/modules/eximport/folder/export_folders.py
@@ -7,11 +7,11 @@ from sqlalchemy.orm import Session
 from core.doc.folder_dto import FolderType
 from core.doc.folder_orm import FolderORM
 from core.project.project_crud import crud_project
+from modules.eximport.export_exceptions import NoDataToExportError
 from modules.eximport.folder.folder_export_schema import (
     FolderExportCollection,
     FolderExportSchema,
 )
-from modules.eximport.no_data_export_error import NoDataToExportError
 from repos.filesystem_repo import FilesystemRepo
