+++ backend/src/modules/eximport/import_exceptions.py
@@ -1,25 +1,16 @@
-from modules.eximport.import_job_dto import (
-    ImportJobType,
-)
+from fastapi import status
+
+from common.exception_handler import exception_handler
+from modules.eximport.import_job_dto import ImportJobType
 
 
+@exception_handler(status.HTTP_500_INTERNAL_SERVER_ERROR)
 class ImportJobPreparationError(Exception):
     def __init__(self, cause: Exception) -> None:
         super().__init__(f"Cannot prepare and create the Import Job! {cause}")
 
 
-class ImportJobAlreadyStartedOrDoneError(Exception):
-    def __init__(self, import_job_id: str) -> None:
-        super().__init__(
-            f"The ImportJob with ID {import_job_id} already started or is done!"
-        )
-
-
-class NoSuchImportJobError(Exception):
-    def __init__(self, import_job_id: str, cause: Exception) -> None:
-        super().__init__(f"There exists not ImportJob with ID {import_job_id}! {cause}")
-
-
+@exception_handler(status.HTTP_500_INTERNAL_SERVER_ERROR)
 class UnsupportedImportJobTypeError(Exception):
     def __init__(self, import_job_type: ImportJobType) -> None:
         super().__init__(f"ImportJobType {import_job_type} is not supported! ")
