--- backend/src/modules/eximport/no_data_export_error.py
@@ -1,3 +1,0 @@
-class NoDataToExportError(Exception):
-    def __init__(self, what_msg: str):
-        super().__init__(what_msg)
