+++ backend/src/modules/eximport/project/export_project.py
@@ -11,9 +11,9 @@ from modules.eximport.bbox_annotations.export_bbox_annotations import (
 )
 from modules.eximport.codes.export_codes import export_all_codes
 from modules.eximport.cota.export_cota import export_all_cota
+from modules.eximport.export_exceptions import NoDataToExportError
 from modules.eximport.folder.export_folders import export_all_folders
 from modules.eximport.memos.export_memos import export_all_memos
-from modules.eximport.no_data_export_error import NoDataToExportError
 from modules.eximport.project_metadata.export_project_metadata import (
     export_all_project_metadatas,
 )
