+++ backend/src/modules/eximport/project_metadata/export_project_metadata.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 
 from core.metadata.project_metadata_orm import ProjectMetadataORM
 from core.project.project_crud import crud_project
-from modules.eximport.no_data_export_error import NoDataToExportError
+from modules.eximport.export_exceptions import NoDataToExportError
 from modules.eximport.project_metadata.project_metadata_export_schema import (
     ProjectMetadataExportCollection,
     ProjectMetadataExportSchema,
