+++ backend/src/repos/elastic/elastic_crud_base.py
@@ -1,8 +1,10 @@
 from typing import Any, Generic, TypeVar
 
 from elasticsearch import Elasticsearch, helpers
+from fastapi import status
 from pydantic import BaseModel
 
+from common.exception_handler import exception_handler
 from repos.elastic.elastic_dto_base import (
     ElasticSearchHit,
     ElasticSearchModelBase,
@@ -16,6 +18,7 @@ UpdateDTOType = TypeVar("UpdateDTOType", bound=BaseModel)
 IndexType = TypeVar("IndexType", bound=IndexBase)
 
 
+@exception_handler(status.HTTP_404_NOT_FOUND)
 class NoSuchObjectInElasticSearchError(Exception):
     def __init__(self, index_name: str, id: int):
         super().__init__(
