+++ ray/pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
     "ffmpeg-python==0.2.0",
     "flash-attn==2.7.4.post1",
     "huggingface_hub==0.29.3",
-    "maverick-coref-de==1.0.5",
+    "maverick-coref-de==1.0.6",
     "omegaconf==2.3.0",
     "quotect==1.1.5",
     "ray[serve]==2.44.1",
